変更 - Prismaに関わる型宣言を移動させる
