+++ app/channels/(types)/index.ts
@@ -7,13 +7,3 @@ export interface ChannelSearchParams {
   play: string[];
   timezone: string[];
 }
-
-export interface PrismaQuery {
-  query: {
-    content: Prisma.ChannelWhereInput;
-    play: Prisma.ChannelWhereInput;
-    timeZone: Prisma.ChannelWhereInput;
-  };
-  year: Prisma.ChannelWhereInput;
-  orderBy: Prisma.SortOrder;
-}
